// webapp/src/api/model/pnPCharacterDTO.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */
import type { PnPCharacterDTOAdvantageTraitsItem } from './pnPCharacterDTOAdvantageTraitsItem';
import type { PnPCharacterDTOCustomFields } from './pnPCharacterDTOCustomFields';
import type { CharacterDescription } from './characterDescription';
import type { PnPCharacterDTODisadvantageTraitsItem } from './pnPCharacterDTODisadvantageTraitsItem';
import type { CharacterEquipment } from './characterEquipment';
import type { CharacterInventory } from './characterInventory';
import type { CharacterLevel } from './characterLevel';
import type { CharacterOrigin } from './characterOrigin';
import type { Spell } from './spell';
import type { CharacterStatsDto } from './characterStatsDto';
import type { PnPCharacterDTOTalents } from './pnPCharacterDTOTalents';

export interface PnPCharacterDTO {
  advantageTraits: PnPCharacterDTOAdvantageTraitsItem[];
  customFields: PnPCharacterDTOCustomFields;
  description: CharacterDescription;
  disadvantageTraits: PnPCharacterDTODisadvantageTraitsItem[];
  equipment: CharacterEquipment;
  id?: string;
  inventory: CharacterInventory;
  level: CharacterLevel;
  origin: CharacterOrigin;
  spells: Spell[];
  stats: CharacterStatsDto;
  talents: PnPCharacterDTOTalents;
}